+++ hbox/hmemory/hfiledescriptor/hfiledescriptor_filev1.h
@@ -21,7 +21,7 @@ struct hfiledescriptor_filev1
     int  (*filev1_write)(hfiledescriptor_fd_t fd,const void *buff,hfiledescriptor_size_t  buff_len);
     int  (*filev1_close)(hfiledescriptor_fd_t fd);
     int  (*filev1_lseek)(hfiledescriptor_fd_t fd, hfiledescriptor_ssize_t offset, int whence);
-    int  (*filev1_ioctl)(hfiledescriptor_fd_t fd, int op, va_list va);
+    int  (*filev1_ioctl)(hfiledescriptor_fd_t fd, unsigned long op, va_list va);
 };
 
 typedef struct hfiledescriptor_filev1 hfiledescriptor_filev1_t;
@@ -34,7 +34,7 @@ int  hfiledescriptor_filev1_read(hfiledescriptor_fd_t fd, void *buff, hfiledescr
 int  hfiledescriptor_filev1_write(hfiledescriptor_fd_t fd,const void *buff,hfiledescriptor_size_t  buff_len);
 int  hfiledescriptor_filev1_close(hfiledescriptor_fd_t fd);
 int  hfiledescriptor_filev1_lseek(hfiledescriptor_fd_t fd, hfiledescriptor_ssize_t offset, int whence);
-int  hfiledescriptor_filev1_ioctl(hfiledescriptor_fd_t fd, int op, va_list va);
+int  hfiledescriptor_filev1_ioctl(hfiledescriptor_fd_t fd, unsigned long op, va_list va);
 
 #ifdef __cplusplus
 }
